final commit

+++ app/controllers/comments_controller.rb
@@ -17,26 +17,34 @@ class CommentsController < ApplicationController
   end
   
   def create
-      post = Post.find(params[:id])
-      comment = post.comment.new(comment_params)
+    post = Post.find(params[:id])
+    comment = post.comments.new(name: params[:name], comment: params[:comment])
   
-      if comment.save
-        redirect_to post_path(post), notice: 'Comment was successfully created.'
-      else
-        redirect_to post_path(post), alert: comment.errors.full_messages.join(', ')
-      end
+    if comment.save
+      render json: comment
+    else
+      render json: {error: comment.errors.full_messages}
+    end
   end
+  
+  
     
 
-	def destroy
-        post = Post.find(params[:id])
-        comment = post.comment.find(params[:id])
-        comment.destroy
-        redirect_to post_path(post), notice: 'Comment was successfully destroyed.'
-      end
+  def destroy
+    post = Post.find(params[:id])
+    comment = post.comments.find_by(id: params[:comment_id])
+    
+    if comment
+      comment.destroy
+      redirect_to post_path(post), notice: 'Comment was successfully destroyed.'
+    else
+      redirect_to post_path(post), alert: 'Comment not found.'
+    end
+  end
+  
       
    
     def comment_params
-        # params.require(:comment).permit(:username, :comment)
+        # params.require(:comment).permit(:name, :comment)
     end
 end
